+++ SnazzCraft/src/snazzcraft-engine/world/world.cpp
@@ -23,7 +23,7 @@ SnazzCraft::World::~World()
 SnazzCraft::World::World(std::string IName, uint32_t ISize, int32_t ISeed)
     : Name(IName), Seed(ISeed), Chunks(std::unordered_map<uint64_t, SnazzCraft::Chunk*>()), VoxelPlacementDisplayPosition(glm::vec3(0.0f))
 {
-    this->Size = (ISize > SnazzCraft::World::MaxSize || ISize <= 0) ? SnazzCraft::World::MaxSize : ISize;
+    this->Size = static_cast<int32_t>((ISize > SnazzCraft::World::MaxSize || ISize <= 0) ? SnazzCraft::World::MaxSize : ISize);
 
     this->WorldHeightMap = new SnazzCraft::HeightMap(static_cast<uint32_t>(this->Size * SnazzCraft::Chunk::Width), -SnazzCraft::Chunk::MaxOceanDepth, SnazzCraft::Chunk::Height - SnazzCraft::Chunk::OceanLevel, this->Seed, 1.0, 0.5, 2.0, 6);
